refactor: remove /Tooltipo folder from Tooltip component namekit-react
