+++ packages/namekit-react/src/client.ts
@@ -1,7 +1,7 @@
 import "@namehash/ens-webfont";
 import "./styles.css";
 
-export { Tooltip } from "./components/Tooltip/Tooltip";
+export { Tooltip } from "./components/Tooltip";
 export {
   CurrencySymbol,
   CurrencySymbolSize,
